This is the stuff I worked on in class. Mostly revamping the physics for the game to be more efficient.

+++ src/Sprite.java
@@ -11,7 +11,7 @@ public class Sprite {
     protected double speedX;
     protected double speedY;
 
-    public Sprite(Image i, int x, int y, int width, int height, double speedX, double speedY) {
+    public Sprite(Image i, double x, double y, int width, int height, double speedX, double speedY) {
         this.i = i;
         this.width = width;
         this.height = height;
